+++ packages/solid/src/utils/getFilteredNames.ts
@@ -1,64 +1,71 @@
 import { untrack } from 'solid-js';
 import { FieldArrayPath, FieldPath, FieldValues, FormState } from '../types';
-import { getFieldArray } from './getFieldArray';
 
 /**
  * Returns a tuple with filtered field and field array names. For each
- * contained array name all fields of the array are added to the field names.
- *
- *  For every field
- * of a field array
+ * specified field array name, the names of the contained fields and field
+ * arrays are also returned. If no name is specified, the name of each field
+ * and field array of the form is returned.
  *
  * @param form The form of the fields and field arrays.
- * @param names The list of field and field array names.
+ * @param arg2 The specified name or names or an object with options.
  *
  * @returns A tuple with the filtered names.
  */
 export function getFilteredNames<
   TFieldValues extends FieldValues,
   TFieldName extends FieldPath<TFieldValues>,
-  TFieldArrayName extends FieldArrayPath<TFieldValues>
+  TFieldArrayName extends FieldArrayPath<TFieldValues>,
+  TOptions extends Record<string, any>
 >(
   form: FormState<TFieldValues>,
-  names: (TFieldName | TFieldArrayName)[]
+  arg2?:
+    | TFieldName
+    | TFieldArrayName
+    | (TFieldName | TFieldArrayName)[]
+    | TOptions
 ): [TFieldName[], TFieldArrayName[]] {
-  // Get all field names of form
-  const allFieldNames = untrack(form.internal.getFieldNames);
+  // Get all field and field array names of form
+  const allFieldNames = untrack(form.internal.getFieldNames) as TFieldName[];
+  const allFieldArrayNames = untrack(
+    form.internal.getFieldArrayNames
+  ) as TFieldArrayName[];
 
-  // Get all field array names of form
-  const allArrayFieldNames = [...form.internal.fieldArrays].map(
-    ([name]) => name
-  );
+  // If names are specified, filter and return them
+  if (typeof arg2 === 'string' || Array.isArray(arg2)) {
+    return (typeof arg2 === 'string' ? [arg2] : arg2)
+      .reduce(
+        (tuple, name) => {
+          // Destructure tuple
+          const [fieldNameSet, fieldArrayNameSet] = tuple;
 
-  // Return tuple with filtered field and field array names
-  return names.reduce(
-    (nameTuple, name) => {
-      // Destructure name tuple
-      const [fieldNames, fieldArrayNames] = nameTuple;
+          // If it is name of a field, add it to field name set
+          if (form.internal.fields.has(name)) {
+            fieldNameSet.add(name as TFieldName);
 
-      // If name is form a field array, add it to field array names
-      if (allArrayFieldNames.includes(name)) {
-        fieldArrayNames.push(name as TFieldArrayName);
-
-        // Get array items to set a listener and be notified when there is a
-        // change when used within an effect
-        getFieldArray(form, name as TFieldArrayName).getItems();
-
-        // Add every name of field of field array to field names
-        allFieldNames.forEach((fieldName) => {
-          if (fieldName.startsWith(name)) {
-            fieldNames.push(fieldName as TFieldName);
+            // If it is name of a field array, add it and name of each field
+            // and field array it contains to field and field array name set
+          } else if (form.internal.fieldArrays.has(name)) {
+            allFieldArrayNames.forEach((fieldArrayName) => {
+              if (fieldArrayName.startsWith(name)) {
+                fieldArrayNameSet.add(fieldArrayName);
+              }
+            });
+            allFieldNames.forEach((fieldName) => {
+              if (fieldName.startsWith(name)) {
+                fieldNameSet.add(fieldName);
+              }
+            });
           }
-        });
 
-        // Otherwise add name to field names
-      } else {
-        fieldNames.push(name as TFieldName);
-      }
+          // Return tuple
+          return tuple;
+        },
+        [new Set(), new Set()] as [Set<TFieldName>, Set<TFieldArrayName>]
+      )
+      .map((set) => [...set]) as [TFieldName[], TFieldArrayName[]];
+  }
 
-      // Return name tuple
-      return nameTuple;
-    },
-    [[], []] as [TFieldName[], TFieldArrayName[]]
-  );
+  // Otherwise return every field and field array name
+  return [allFieldNames, allFieldArrayNames];
 }